requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.9.0
psycopg2==2.6.1
SQLAlchemy==1.0.7
Werkzeug==0.10.4
